# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Campus Coordinator API",
    "description": "This is an API server for the Campus Coordinator API project",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/room": {
      "get": {
        "tags": [
          "Rooms"
        ],
        "summary": "Get all rooms",
        "responses": {
          "200": {
            "description": "List of room",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Room"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Rooms"
        ],
        "summary": "Create a new room",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Room"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Room created successfully"
          }
        }
      }
    },
    "/room/{id}": {
      "get": {
        "tags": [
          "Rooms"
        ],
        "summary": "Get a room by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Room data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          },
          "404": {
            "description": "Room not found"
          }
        }
      },
      "put": {
        "tags": [
          "Rooms"
        ],
        "summary": "Update a room by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Room"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Room updated successfully"
          },
          "404": {
            "description": "Room not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Rooms"
        ],
        "summary": "Delete a room by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Room deleted successfully"
          },
          "404": {
            "description": "Room not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Room": {
        "type": "object",
        "properties": {
          "buildingName": {
            "type": "string",
            "example": "Art Center"
          },
          "roomNumber": {
            "type": "string",
            "example": "303"
          },
          "capacity": {
            "type": "integer",
            "example": 25
          },
          "students": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "664b8df7254bd4e3f9a7f9b31"
            }
          }
        },
        "required": [
          "buildingName",
          "roomNumber",
          "capacity"
        ]
      }
    }
  }
}